Support Safello redirect URLs
When the user buys bitcoin through Safello, they will be redirected to
the Yoti app to verify their identity.

Previously, when the process was done, they
would be redirected to a browser window to continue their purchase, but
since bluewallet has a `bluewallet://` redirect URI associated to its
app ID, we can redirect them back to the app with a
`safello-state-token` query parameter appended to the URI instead.

Bluewallet then simply needs to check if that query parameter is
present, and if so, open a new WebView with the same Safello URL +
`stateToken=xxx` in order to resume the user's session. This way, the
whole buying process can take place inside Bluewallet's app, which
is a clear improvement in terms of user experience.