REF: Correcting Punctuation Marks
Periods, commas, dashes (hyphens and em dashes).

Notable Issues
1. No comma is required after “please.”
2. A period is required to end a sentence. A full sentence must have a full stop at the end.
3. Incorrect use of the exclamation mark. Here is a quick read: grammarly.com/blog/exclamation-mark.
4. When a sentence is followed by a dependent phrase, such as the one below, an em dash (—) is recommended instead of a comma:
Example: This is your transaction's hex—signed and ready to be broadcasted to the network. (A dependent phrases such as “signed and ready to be broadcasted to the network” cannot stand on its own as a full sentence because it is not.)
5. Extra space between the last word of the sentence and period, namely: “…has been saved in your Downloads folder .” The extra space must be omitted.

There is a reason for each specific change that I made, thus making all of the necessary. I would be happy to provide sources if necessary.